Исправления по замечаниям
