+++ src/components/film-card/templates/film-card.js
@@ -1,7 +1,4 @@
-import AbstractComponent from './abstract-component.js';
-
-const createFilmCardTemplate = (film) => {
-  const {title, rating, year, duration, genre, poster, description, commentsCount} = film;
+const createFilmCardTemplate = ({title, rating, year, duration, genre, poster, description, commentsCount}) => {
 
   return (
     `<article class="film-card">
@@ -24,29 +21,4 @@ const createFilmCardTemplate = (film) => {
   );
 };
 
-export default class FilmCard extends AbstractComponent {
-  constructor(film) {
-    super();
-
-    this._film = film;
-  }
-
-  getTemplate() {
-    return createFilmCardTemplate(this._film);
-  }
-
-  setFilmCardTitleHandler(handler) {
-    this.getElement().querySelector(`.film-card__title`)
-      .addEventListener(`click`, handler);
-  }
-
-  setFilmCardPosterHandler(handler) {
-    this.getElement().querySelector(`.film-card__poster`)
-      .addEventListener(`click`, handler);
-  }
-
-  setFilmCardCommentsHandler(handler) {
-    this.getElement().querySelector(`.film-card__comments`)
-      .addEventListener(`click`, handler);
-  }
-}
+export {createFilmCardTemplate};
