--- src/components/films-block.js
@@ -1,13 +1,0 @@
-import AbstractComponent from './abstract-component.js';
-
-const createFilmsBlockTemplate = () => {
-  return (
-    `<section class="films"></section>`
-  );
-};
-
-export default class FilmsBlock extends AbstractComponent {
-  getTemplate() {
-    return createFilmsBlockTemplate();
-  }
-}
