+++ src/components/films-block/templates/films-block.js
@@ -1,0 +1,7 @@
+const createFilmsBlockTemplate = () => {
+  return (
+    `<section class="films"></section>`
+  );
+};
+
+export {createFilmsBlockTemplate};
