+++ src/components/films-list/films-list.js
@@ -1,0 +1,8 @@
+import AbstractComponent from '../abstract-component.js';
+import {createFilmsListTemplate} from './templates/films-list.js';
+
+export default class FilmsList extends AbstractComponent {
+  getTemplate() {
+    return createFilmsListTemplate();
+  }
+}
