--- src/components/no-films.js
@@ -1,13 +1,0 @@
-import AbstractComponent from './abstract-component.js';
-
-const createNoFilmsTemplate = () => {
-  return (
-    `<h2 class="films-list__title">There are no movies in our database</h2>`
-  );
-};
-
-export default class ShowMoreButton extends AbstractComponent {
-  getTemplate() {
-    return createNoFilmsTemplate();
-  }
-}
