+++ src/components/no-films/no-films.js
@@ -1,0 +1,8 @@
+import AbstractComponent from '../abstract-component.js';
+import {createNoFilmsTemplate} from './templates/no-films.js';
+
+export default class ShowMoreButton extends AbstractComponent {
+  getTemplate() {
+    return createNoFilmsTemplate();
+  }
+}
