+++ src/components/no-films/templates/no-films.js
@@ -1,0 +1,7 @@
+const createNoFilmsTemplate = () => {
+  return (
+    `<h2 class="films-list__title">There are no movies in our database</h2>`
+  );
+};
+
+export {createNoFilmsTemplate};
