+++ src/main.js
@@ -1,15 +1,15 @@
-import HeaderProfileComponent from './components/header-profile.js';
-import SiteMenuComponent from './components/site-menu.js';
+import HeaderProfileComponent from './components/header-profile/header-profile.js';
+import SiteMenuComponent from './components/site-menu/site-menu.js';
+import SortComponent from './components/sort/sort.js';
 import PageController from "./controllers/page.js";
-import {generateMenuItems} from './mock/menu.js';
 import {render, RenderPosition} from './utils/render.js';
 
 const siteHeaderElement = document.querySelector(`.header`);
 render(siteHeaderElement, new HeaderProfileComponent(), RenderPosition.BEFOREEND);
 
 const siteMainElement = document.querySelector(`.main`);
-const menuItems = generateMenuItems();
-render(siteMainElement, new SiteMenuComponent(menuItems), RenderPosition.BEFOREEND);
+render(siteMainElement, new SiteMenuComponent(), RenderPosition.BEFOREEND);
+render(siteMainElement, new SortComponent(), RenderPosition.BEFOREEND);
 
 const pageController = new PageController();
 pageController.render();
